backend/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
defusedxml==0.7.1
Django==4.2
django-colorfield==0.8.0
django-cors-headers==3.14.0
django-filter==23.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
Pillow==9.5.0
pip==21.2.4
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.6.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
reportlab==3.6.12
requests==2.28.2
requests-oauthlib==1.3.1
setuptools==58.1.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.1
sqlparse==0.4.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
webcolors==1.13